Add Node and Box classes
Modified the internal implementation of dots and boxes to have Node
an a Box structures. A node represents the dots which are to be
connected by the players. A box is the structure with four nodes as
its corners and is rewarded to a player once its four sides are drawn.